Rename folder stencil to masking
