+++ str2d/__init__.py
@@ -1,9 +1,9 @@
 from str2d.str2d import Str2D
 from str2d.str2d import utils
 
-# will need (I think) to add the following, once mask/stencil is extracted,
+# will need (I think) to add the following, once mask/masking is extracted,
 # in order to maintain the appearance of a  flat import structure for clients
-# from stencil.stencil import Mask
-# from stencil.stencil import Perforation        #<-- maybe not needed
-# from stencil.stencil import SOLID, PUNCHED     #<-- maybe not needed
-# from stencil.stencil import Stencil
+# from masking.masking import Mask
+# from masking.masking import Perforation        #<-- maybe not needed
+# from masking.masking import SOLID, PUNCHED     #<-- maybe not needed
+# from masking.masking import Stencil
